trees/0105_construct_binary_tree_from_preorder_and_inorder_traversal/solution_test.go
package constructbinarytreefrompreorderandinordertraversal105

import (
	"reflect"
	"testing"

	. "github.com/aliml92/leetcode/datastructure"
)

func Test_buildTree(t *testing.T) {
	type args struct {
		preorder []int
		inorder  []int
	}
	tests := []struct {
		name string
		args args
		want *TreeNode
	}{
		{
			"example1",
			args{
				[]int{3, 9, 20, 15, 7},
				[]int{9, 3, 15, 20, 7},
			},
			BuildTree([]interface{}{3, 9, 20, nil, nil, 15, 7}),
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := buildTree(tt.args.preorder, tt.args.inorder); !reflect.DeepEqual(got, tt.want) {
				t.Errorf("buildTree() = %v, want %v", got, tt.want)
			}
		})
	}
}
